# docs/api/openapi.yaml
openapi: 3.1.0
servers:
- url: https://dev-ddan-ddan-api.ddmz.org/
  description: Development API server URL
- url: https://ddan-ddan-api.ddmz.org/
  description: Production API server URL
- url: http://localhost:8080/
  description: Local API server URL
security:
- JWT: []
...
paths:
...
  /v1/admin/pet-catalog/{type}:
    put:
      tags:
      - Admin · 펫 카탈로그
      summary: 펫 카탈로그 수정
      operationId: update
      parameters:
      - name: type
        in: path
        description: 수정할 펫 카탈로그 타입
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetCatalogAdminUpdateRequest"
        required: true
      responses:
        "200":
          description: 수정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetCatalogAdminItemResponse"
        "400":
          description: 요청 값 검증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 관리자 권한 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 펫 카탈로그 항목 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
      - Admin · 펫 카탈로그
      summary: 펫 카탈로그 비활성화 (soft delete)
      operationId: softDelete
      parameters:
      - name: type
        in: path
        description: 비활성화할 펫 카탈로그 타입
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 비활성화 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetCatalogAdminItemResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 관리자 권한 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 펫 카탈로그 항목 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/admin/pet-catalog:
    get:
      tags:
      - Admin · 펫 카탈로그
      summary: 전체 펫 카탈로그 조회 (비활성 포함)
      operationId: list
      responses:
        "200":
          description: 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetCatalogAdminListResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 관리자 권한 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
      - Admin · 펫 카탈로그
      summary: 펫 카탈로그 신규 등록
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetCatalogAdminCreateRequest"
        required: true
      responses:
        "200":
          description: 등록 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetCatalogAdminItemResponse"
        "400":
          description: 요청 값 검증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 관리자 권한 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 중복된 펫 카탈로그 타입
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/pets/{petId}:
    get:
      tags:
      - 펫
      summary: 펫 조회
      description: 특정 펫을 조회합니다.
      operationId: getPet
      parameters:
      - name: petId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PetResponse"
...
components:
  schemas:
...
    PetCatalogAdminUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        colorCode:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        isActive:
          type: boolean
        displayOrder:
          type: integer
          format: int32
          minimum: 0
        levels:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PetCatalogLevelRequest"
          minProperties: 1
      required:
      - colorCode
      - displayOrder
      - isActive
      - levels
      - name
    PetCatalogLevelRequest:
      type: object
      properties:
        imageUrl:
          type: string
          minLength: 1
        lottieDefaultUrl:
          type: string
          minLength: 1
        lottiePlayEatUrl:
          type: string
          minLength: 1
      required:
      - imageUrl
      - lottieDefaultUrl
      - lottiePlayEatUrl
    ErrorResponse:
      type: object
      description: 에러 응답 DTO
      properties:
        code:
          type: string
          description: 에러 응답 코드
        message:
          type: string
          description: 에러 응답 메시지
        data:
          description: 추가 데이터
      required:
      - code
      - message
    PetCatalogAdminItemResponse:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        colorCode:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
          format: int32
        levels:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PetCatalogLevelResponse"
      required:
      - colorCode
      - displayOrder
      - isActive
      - levels
      - name
      - type
    PetCatalogLevelResponse:
      type: object
      properties:
        imageUrl:
          type: string
        lottieDefaultUrl:
          type: string
        lottiePlayEatUrl:
          type: string
      required:
      - imageUrl
      - lottieDefaultUrl
      - lottiePlayEatUrl
...
    PetResponse:
      type: object
      description: 펫 정보 응답 DTO
      properties:
        id:
          type: string
          description: Pet 식별 ID
          example: ABCDEF1234567
        type:
          type: string
          description: Pet 종류 (PetCatalog.type)
          example: DOG
        level:
          type: integer
          format: int32
          description: Pet 레벨
          example: 1
        expPercent:
          type: number
          format: double
          description: 현재 레벨에 해당하는 경험치 퍼센트
          example: 15.5
      required:
      - expPercent
      - id
      - level
      - type
...
    PetCatalogAdminCreateRequest:
      type: object
      properties:
        type:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        colorCode:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        isActive:
          type: boolean
        displayOrder:
          type: integer
          format: int32
          minimum: 0
        levels:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PetCatalogLevelRequest"
          minProperties: 1
      required:
      - colorCode
      - displayOrder
      - isActive
      - levels
      - name
      - type
...
    PetCatalogAdminListResponse:
      type: object
      properties:
        pets:
          type: array
          items:
            $ref: "#/components/schemas/PetCatalogAdminItemResponse"
      required:
      - pets
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT